feat: implement data retrieval functions for manufacturers, categories, and conditions using Prisma
