+++ app/api/v2/barometers/[slug]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { type PrismaClient } from '@prisma/client'
+import { getBarometer } from './getters'
 import { getPrismaClient } from '@/prisma/prismaClient'
 
 interface Params {
@@ -7,49 +7,6 @@ interface Params {
     slug: string
   }
 }
-
-async function getBarometer(prisma: PrismaClient, slug: string) {
-  const barometer = await prisma.barometer.findFirstOrThrow({
-    where: {
-      slug: {
-        equals: slug,
-        mode: 'insensitive',
-      },
-    },
-    include: {
-      category: true,
-      condition: {
-        select: {
-          id: true,
-          name: true,
-          description: true,
-          value: true,
-        },
-      },
-      manufacturer: {
-        select: {
-          id: true,
-          city: true,
-          name: true,
-          country: true,
-          description: true,
-        },
-      },
-      images: {
-        orderBy: {
-          order: 'asc',
-        },
-      },
-    },
-  })
-  return {
-    ...barometer,
-    // temporarily
-    images: barometer.images.map(img => img.url),
-  }
-}
-
-export type BarometerDTO = Awaited<ReturnType<typeof getBarometer>>
 
 /**
  * Get Barometer details by slug
