+++ app/api/v2/categories/route.ts
@@ -1,27 +1,6 @@
 import { NextResponse } from 'next/server'
-import { type PrismaClient } from '@prisma/client'
 import { getPrismaClient } from '@/prisma/prismaClient'
-
-async function getCategories(prisma: PrismaClient) {
-  return prisma.category.findMany({
-    orderBy: {
-      order: 'asc',
-    },
-    select: {
-      id: true,
-      name: true,
-      label: true,
-      order: true,
-      image: {
-        select: {
-          url: true,
-        },
-      },
-    },
-  })
-}
-
-export type CategoryListDTO = Awaited<ReturnType<typeof getCategories>>
+import { getCategories } from './getters'
 
 /**
  * Get Categories list
