+++ app/api/v2/conditions/getters.ts
@@ -1,0 +1,17 @@
+import { PrismaClient } from '@prisma/client'
+
+export async function getConditions(prisma: PrismaClient) {
+  return prisma.condition.findMany({
+    orderBy: {
+      value: 'asc',
+    },
+    select: {
+      id: true,
+      name: true,
+      value: true,
+      description: true,
+    },
+  })
+}
+
+export type ConditionListDTO = Awaited<ReturnType<typeof getConditions>>
