+++ app/api/v2/manufacturers/[id]/getters.ts
@@ -1,0 +1,10 @@
+import { PrismaClient } from '@prisma/client'
+
+export async function getManufacturer(prisma: PrismaClient, id: string) {
+  return prisma.manufacturer.findUnique({
+    where: {
+      id,
+    },
+  })
+}
+export type ManufacturerDTO = Awaited<ReturnType<typeof getManufacturer>>
