+++ app/types.ts
@@ -1,11 +1,11 @@
-export type { BarometerListDTO, ParameterizedBarometerListDTO } from './api/v2/barometers/route'
-export type { CategoryListDTO } from './api/v2/categories/route'
-export type { CategoryDTO } from './api/v2/categories/[name]/route'
-export type { BarometerDTO } from './api/v2/barometers/[slug]/route'
-export type { ConditionListDTO } from './api/v2/conditions/route'
-export type { ManufacturerListDTO } from './api/v2/manufacturers/route'
-export type { ManufacturerDTO } from './api/v2/manufacturers/[id]/route'
-export type { SearchResultsDTO } from './api/v2/search/route'
+export type { BarometerListDTO, ParameterizedBarometerListDTO } from './api/v2/barometers/getters'
+export type { CategoryListDTO } from './api/v2/categories/getters'
+export type { CategoryDTO } from './api/v2/categories/[name]/getters'
+export type { BarometerDTO } from './api/v2/barometers/[slug]/getters'
+export type { ConditionListDTO } from './api/v2/conditions/getters'
+export type { ManufacturerListDTO } from './api/v2/manufacturers/getters'
+export type { ManufacturerDTO } from './api/v2/manufacturers/[id]/getters'
+export type { SearchResultsDTO } from './api/v2/search/search'
 /**
  * Barometer dimensions database JSON structure
  */
